refactor: repackage auth
