+++ module-core/src/main/java/com/codingbottle/auth/entity/Role.java
@@ -1,4 +1,4 @@
-package com.codingbottle.auth.domain.entity;
+package com.codingbottle.auth.entity;
 
 public enum Role {
     ROLE_ADMIN, ROLE_USER
